création de la page du catalogue
