+++ src/views/HomeView.vue
@@ -1,81 +1,7 @@
-<script setup lang="ts">
-import ListCard from '@/components/ListCard.vue'
-
-const listTestRoom = [
-  {
-    name: 'Pièce n°1',
-    picture: 'room1_small',
-    adress: '20 rue de la joie, 67100 Strasbourg',
-    groupe: 'Réunion',
-    capacite: 10,
-    superficie: '30x50',
-    equipements: [
-      { name: 'chaise', nbr: 10 },
-      { name: 'tableau blanc', nbr: 1 },
-      { name: 'vidéo projecteur', nbr: 1 },
-    ],
-  },
-  {
-    name: 'Pièce n°2',
-    picture: 'room2_small',
-    adress: '20 rue de la joie, 67100 Strasbourg',
-    groupe: 'Détente',
-    capacite: 10,
-    superficie: '30x50',
-    equipements: [
-      { name: 'chaise', nbr: 10 },
-      { name: 'tableau blanc', nbr: 1 },
-      { name: 'vidéo projecteur', nbr: 1 },
-    ],
-  },
-  {
-    name: 'Pièce n°3',
-    picture: 'room3_small',
-    adress: '20 rue de la joie, 67100 Strasbourg',
-    groupe: 'Réunion',
-    capacite: 10,
-    superficie: '30x50',
-    equipements: [
-      { name: 'chaise', nbr: 10 },
-      { name: 'tableau blanc', nbr: 1 },
-      { name: 'vidéo projecteur', nbr: 1 },
-    ],
-  },
-  {
-    name: 'Pièce n°4',
-    picture: 'room4_small',
-    adress: '20 rue de la joie, 67100 Strasbourg',
-    groupe: 'Détente',
-    capacite: 10,
-    superficie: '30x50',
-    equipements: [
-      { name: 'chaise', nbr: 10 },
-      { name: 'tableau blanc', nbr: 1 },
-      { name: 'vidéo projecteur', nbr: 1 },
-    ],
-  },
-  {
-    name: 'Pièce n°5',
-    picture: 'room5_small',
-    adress: '20 rue de la joie, 67100 Strasbourg',
-    groupe: 'Réunion',
-    capacite: 10,
-    superficie: '30x50',
-    equipements: [
-      { name: 'chaise', nbr: 10 },
-      { name: 'tableau blanc', nbr: 1 },
-      { name: 'vidéo projecteur', nbr: 1 },
-    ],
-  },
-]
-</script>
+<script setup lang="ts"></script>
 
 <template>
-  <main>
-    <div>
-      <p>Accueil View</p>
-
-      <ListCard :listRoom="listTestRoom" />
-    </div>
+  <main class="h-[200vh]">
+    <h1>Accueil</h1>
   </main>
 </template>
